added validations to wuth and separatd into different components
